requirements.txt

apache-airflow==2.2.5
pandas
SQLAlchemy
